# Conflicts:
#	src/main/web/WEB-INF/product/list.jsp
#	src/main/web/admin/add_product.jsp

+++ src/main/java/action/AddressAction.java
@@ -1,6 +1,7 @@
 package action;
 
 import com.opensymphony.xwork2.ActionSupport;
+import constant.Key;
 import org.hibernate.Hibernate;
 import org.hibernate.HibernateException;
 import po.Address;
@@ -8,7 +9,9 @@ import po.User;
 import service.AddressService;
 import service.UserService;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Created by zhang on 2016/6/27.
@@ -29,6 +32,7 @@ public class AddressAction extends ActionSupport {
     private String addressDetail;
     private Integer defaultAddressId;
     private int add;
+    private Map<String, Object> data = new HashMap<>();
 
     public String viewAddress() {
         try {
@@ -36,8 +40,7 @@ public class AddressAction extends ActionSupport {
             addressList = user.getAddresses();
             defaultAddressId = user.getDefaultAddressId();
             return SUCCESS;
-        }
-        catch (HibernateException e){
+        } catch (HibernateException e) {
             e.printStackTrace();
             return ERROR;
         }
@@ -63,19 +66,30 @@ public class AddressAction extends ActionSupport {
                 userService.update(user);
             }
             return SUCCESS;
-        }
-        catch(HibernateException e){
+        } catch (HibernateException e) {
             e.printStackTrace();
             return ERROR;
         }
     }
 
+
+    public void validateAddAddress() {
+        if (!mobile.matches("^\\w{11}$")) {
+            addFieldError("mobile", "电话号码请输入11位数字");
+            return;
+        }
+        super.validate();
+    }
+
     public String deleteAddress() {
         try {
+            Address address = addressService.get(addressId);
+            if (address == null) {
+                return ERROR;
+            }
             addressService.deleteAddress(addressId);
             return SUCCESS;
-        }
-        catch(HibernateException e){
+        } catch (HibernateException e) {
             e.printStackTrace();
             return ERROR;
         }
@@ -97,24 +111,24 @@ public class AddressAction extends ActionSupport {
 
             addressService.update(address);
             return SUCCESS;
-        }
-        catch (HibernateException e){
+        } catch (HibernateException e) {
             e.printStackTrace();
             return ERROR;
         }
     }
 
-    public String setDefaultAddress(){
+    public String setDefaultAddress() {
         try {
             user = userService.getCurrentUser();
             user.setDefaultAddressId(defaultAddressId);
             userService.update(user);
-            return SUCCESS;
-        }
-        catch (HibernateException e){
+            data.put(Key.RESULT, SUCCESS);
+        } catch (HibernateException e) {
             e.printStackTrace();
-            return ERROR;
+            data.put(Key.RESULT, ERROR);
         }
+        return SUCCESS;
+
     }
 
     public UserService getUserService() {
@@ -243,5 +257,13 @@ public class AddressAction extends ActionSupport {
 
     public void setAdd(int add) {
         this.add = add;
+    }
+
+    public Map<String, Object> getData() {
+        return data;
+    }
+
+    public void setData(Map<String, Object> data) {
+        this.data = data;
     }
 }
